+++ dbus-serialbattery/dbus-serialbattery.py
@@ -427,6 +427,37 @@ def main():
             can_thread.setup_can(channel=port, bitrate=busspeed, force=True)
             sleep(2)
 
+    # MQTT
+    elif port == "mqtt":
+        """
+        Import MQTT class only if it's a MQTT connection; otherwise, the driver won't start due to missing Python modules.
+        This prevents issues when using the driver exclusively with a serial connection.
+        """
+
+        if len(sys.argv) <= 2:
+            logger.error(">>> MQTT topic is missing in the command line arguments")
+            sleep(60)
+            exit_driver(None, None, 1)
+        else:
+            from bms.generic_mqtt import Generic_Mqtt
+
+            # TODO: Currently only one topic is supported
+            # TODO: No timeout implemented yet
+
+            mqtt_topic = sys.argv[2]
+
+            # Split mqtt topics by comma to allow multiple batteries
+            # TODO: implement multiple battery support
+            mqtt_topics = mqtt_topic.split(",")
+            logger.info("MQTT topics: " + ", ".join(mqtt_topics))
+
+            # do not remove mqtt_ prefix, since the dbus service cannot be only numbers
+            testbms = Generic_Mqtt("mqtt_0", None, mqtt_topics[0])
+
+            if testbms.test_connection():
+                logger.info("-- Connection established to " + testbms.__class__.__name__)
+                battery[0] = testbms
+
     # SERIAL
     else:
         # check if BMS_TYPE is not empty and all BMS types in the list are supported
@@ -487,21 +518,22 @@ def main():
     # get first key from battery dict
     first_key = list(battery.keys())[0]
 
-    # try using active callback on this battery (normally only used for Bluetooth BMS)
-    if not battery[first_key].use_callback(lambda: poll_battery(mainloop)):
-        # change poll interval if set in config
+    # check if active callback is enabled for this battery
+    # normally used only for BLE and MQTT batteries
+    use_active_callback = battery[first_key].use_callback(lambda: poll_battery(mainloop))
+    if use_active_callback:
+        logger.info("Polling interval: active callback used")
+    else:
+        # set poll interval from config if provided
         if POLL_INTERVAL is not None:
             battery[first_key].poll_interval = POLL_INTERVAL
 
         logger.info(f"Polling interval: {battery[first_key].poll_interval/1000:.3f} s")
-
-        # if not possible, poll the battery every poll_interval milliseconds
+        # schedule periodic polling
         gobject.timeout_add(
             battery[first_key].poll_interval,
             lambda: poll_battery(mainloop),
         )
-    else:
-        logger.info("Polling interval: active callback used")
 
     # print log at this point, else not all data is correctly populated
     for key_address in battery: